Che6.0

+++ wsagent/che-core-api-project-shared/src/main/java/org/eclipse/che/api/project/shared/dto/ProjectSearchResponseDto.java
@@ -19,7 +19,7 @@ public interface ProjectSearchResponseDto {
 
   ProjectSearchResponseDto withItemReferences(List<SearchResultDto> itemReferences);
 
-  int getTotalHits();
+  long getTotalHits();
 
-  ProjectSearchResponseDto withTotalHits(int totalHits);
+  ProjectSearchResponseDto withTotalHits(long totalHits);
 }
